Change Buffer.size to ULong

+++ backend/mock/src/commonMain/kotlin/io/technoirlab/rhi/mock/MockBuffer.kt
@@ -4,10 +4,10 @@ import io.technoirlab.rhi.core.Buffer
 import kotlinx.io.Source
 
 internal abstract class MockBuffer(
-    override val size: UInt
+    override val size: ULong
 ) : Buffer {
 
-    override fun updateData(source: Source, size: UInt, offset: UInt) = Unit
+    override fun updateData(source: Source, size: ULong, offset: ULong) = Unit
 
     override fun close() = Unit
 }
