+++ backend/mock/src/commonMain/kotlin/io/technoirlab/rhi/mock/geometry/MockIndexBuffer.kt
@@ -5,7 +5,7 @@ import io.technoirlab.rhi.core.geometry.IndexType
 import io.technoirlab.rhi.mock.MockBuffer
 
 internal class MockIndexBuffer(
-    size: UInt,
+    size: ULong,
     override val indexCount: UInt,
     override val indexType: IndexType
 ) : MockBuffer(size), IndexBuffer
